feat: add local search service (#186)
* feat: search API CLI command


* feat: extension, locator queries revamp


* fix: grpc numbering


* fix: cr remarks


* fix: cr remarks


* fix: cmd args


* fix: cr remarks


* fix: cr remarks


* fix: cr remarks


* fix: cr remarks


---------

+++ server/search/sqlite/skill.go
@@ -13,11 +13,11 @@ import (
 type Skill struct {
 	gorm.Model
 	AgentID uint   `gorm:"not null;index"`
-	SkillID uint32 `gorm:"not null"`
+	SkillID uint64 `gorm:"not null"`
 	Name    string `gorm:"not null"`
 }
 
-func (skill *Skill) GetID() uint32 {
+func (skill *Skill) GetID() uint64 {
 	return skill.SkillID
 }
